% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_Rcplex.R
\name{constraint_caliper}
\alias{constraint_caliper}
\title{Define a distance caliper constraint}
\usage{
constraint_caliper(dist_matrix, caliper, D = NULL)
}
\arguments{
\item{dist_matrix}{EITEHR a N x N matrix of distances between units. Or a Nt x Nc matrix of distances between treated and control units. If the former the the argument D must also be provided.}

\item{caliper}{either: a Nt x Nc matrix of maximum allowed distances between treated and control units or a scalar denoting the max allowed distance between any pair of units.}

\item{D}{binary treatment indicator vector if not left NULL, then dist_matrix is assumed to be NxN}
}
\value{
A function to be used in conjunction with one of the MB computations
}
\description{
Define a distance caliper constraint
}
