% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_Rcplex.R
\name{constraint_moment}
\alias{constraint_moment}
\title{Defines a moment distance constraint}
\usage{
constraint_moment(X, m, mom, eps)
}
\arguments{
\item{X}{matrix of matching covariates.}

\item{m}{Integer. total number of matches to be made.}

\item{mom}{Integer. moment to constrain.}

\item{eps}{Either a vector of numeric tolerances with length(eps) == ncol(X) or a scalar defining a global tolerance for all covariates.}
}
\value{
A function to be used in conjunction with one of the MB computations
}
\description{
Defines a moment distance constraint
}
