% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_Rcplex.R
\name{constraint_quantile}
\alias{constraint_quantile}
\title{Defines a quantile distance constraint}
\usage{
constraint_quantile(X, D, m, quantiles, eps)
}
\arguments{
\item{X}{matrix of matching covariates.}

\item{D}{Binary treatment indicator vector}

\item{m}{Integer. total number of matches to be made.}

\item{quantiles}{List. A list of vectors with length(quantiles) == ncol(X). Each vector should specify the values of the Nq quantiles of each covariate that the user wish to constrain.}

\item{eps}{Either a vector of numeric tolerances with as many elements as the total elements in quantiles or a scalar defining a global tolerance for all quantiles of all covariates.}
}
\value{
A function to be used in conjunction with one of the MB computations
}
\description{
Defines a quantile distance constraint
}
