% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_Rcplex.R
\name{progressive_constraint_moment}
\alias{progressive_constraint_moment}
\title{Defines a progressively-relaxed moment distance constraint}
\usage{
progressive_constraint_moment(
  X,
  D,
  m,
  mom,
  eps,
  init_rf = 0,
  end_rf = NULL,
  rf_step = NULL,
  nsteps = NULL
)
}
\arguments{
\item{X}{matrix of matching covariates.}

\item{D}{Binary treatment indicator vector}

\item{m}{Integer. total number of matches to be made.}

\item{mom}{Integer. moment to constrain.}

\item{eps}{Either a vector of numeric tolerances with length(eps) == ncol(X) or a scalar defining a global tolerance for all covariates.}

\item{init_rf}{The initial relaxation factor (a proportion > 0) of the constraints}

\item{end_rf}{The final relaxation factor that should be reached from the initial in nsteps steps. If provided then rf_step or nsteps can be automatically computed depending on which of the two is provided.}

\item{rf_step}{The amount to increase the relaxation factor by each step. If provided then either end_rf or nsteps can be automatically computed depending on which of the two is provided.}

\item{nsteps}{The total number of times MBs should be estimated. f provided then either end_rf or rf_step can be automatically computed depending on which of the two is provided.}
}
\value{
A function to be used in conjunction with one of the MB computations
}
\description{
Defines a progressively-relaxed moment distance constraint
}
