% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_bounds.R
\name{summary.mb_results}
\alias{summary.mb_results}
\title{Summary function for matching bounds results}
\usage{
\method{summary}{mb_results}(object, ...)
}
\description{
Summary function for matching bounds results
}
