% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_bounds.R
\name{summary.prog_mb_results}
\alias{summary.prog_mb_results}
\title{Summary function for progressive matching bounds results}
\usage{
\method{summary}{prog_mb_results}(object, ...)
}
\description{
Summary function for progressive matching bounds results
}
