% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_Rcplex.R
\name{constraint_caliper}
\alias{constraint_caliper}
\title{Define a distance caliper constraint}
\usage{
constraint_caliper(dist_matrix, caliper)
}
\arguments{
\item{dist_matrix}{a N x N matrix of distances between units.}

\item{caliper}{either: a Nt x Nc matrix of maximum allowed distances between treated and control units or a scalar denoting the max allowed distance between any pair of units.}
}
\value{
A function to be used in conjunction with one of the MB computations
}
\description{
Define a distance caliper constraint
}
